requirements.txt
absl-py==2.2.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asciitree==0.3.3
astor==0.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
clearml==1.18.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.0
cucim-cu12==24.8.0
cupy-cuda12x==13.4.1
cycler==0.12.1
databricks-sdk==0.49.0
Deprecated==1.2.18
docker==7.1.0
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
fasteners==0.19
fastrlock==0.8.3
filelock==3.11.0
fire==0.7.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.1
furl==2.1.4
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
json-tricks==3.17.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.6.2
lpips==0.1.4
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mlflow==2.21.2
mlflow-skinny==2.21.2
monai==1.4.0
mpmath==1.3.0
networkx==3.2.1
nibabel==5.2.1
ninja==1.11.1.4
nni==3.0
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
openslide-python==1.4.1
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
optuna==4.2.1
orderedmultidict==1.0.1
packaging==24.2
pandas==2.2.3
pathlib2==2.3.7.post1
pillow==11.1.0
prettytable==3.16.0
protobuf==4.25.6
psutil==7.0.0
pyamg==5.2.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pydicom==2.4.4
PyJWT==2.9.0
pynrrd==1.1.3
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.11
pytz==2025.2
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
responses==0.25.7
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
schema==0.7.7
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.24.1
setproctitle==1.3.5
SimpleITK==2.4.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==1.0.15
tokenizers==0.15.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.38.1
triton==3.2.0
typeguard==4.1.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.16.3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
zarr==2.18.2
zipp==3.21.0
